Controladores nuevos: Song, Comment y User.
